requirements.txt
flask
transformers
pdfminer.six
python-pptx
PyPDF2